Remove stdenv hash finding from massrebuilder
